requirements.txt
fastapi==0.111.0
uvicorn==0.30.1

langchain==0.2.7
langchain-community==0.2.7

faiss-cpu==1.8.0
sentence-transformers==2.7.0
pypdf==4.2.0
python-dotenv==1.0.1
